+++ internal/provider/resource_tfe_project.go
@@ -55,6 +55,19 @@ func resourceTFEProject() *schema.Resource {
 				Computed: true,
 				ForceNew: true,
 			},
+
+			"tags": {
+				Type:     schema.TypeMap,
+				Optional: true,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+				},
+			},
+
+			"ignore_additional_tags": {
+				Type:     schema.TypeBool,
+				Optional: true,
+			},
 		},
 	}
 }
@@ -98,9 +111,28 @@ func resourceTFEProjectRead(ctx context.Context, d *schema.ResourceData, meta in
 		return diag.FromErr(err)
 	}
 
+	tagBindings, err := config.Client.Projects.ListTagBindings(ctx, project.ID)
+	if err != nil && !errors.Is(err, tfe.ErrResourceNotFound) {
+		return diag.FromErr(err)
+	}
+	if err != nil {
+		log.Printf("[DEBUG] Project %s no longer exists or tag bindings are not supported by this instance", d.Id())
+		tagBindings = []*tfe.TagBinding{}
+	}
+
+	bindings := make(map[string]interface{})
+	configBindings := d.Get("tags").(map[string]interface{})
+	for _, binding := range tagBindings {
+		_, ok := configBindings[binding.Key]
+		if ok || !d.Get("ignore_additional_tags").(bool) {
+			bindings[binding.Key] = binding.Value
+		}
+	}
+
 	d.Set("name", project.Name)
 	d.Set("description", project.Description)
 	d.Set("organization", project.Organization.Name)
+	d.Set("tags", bindings)
 
 	return nil
 }
@@ -111,6 +143,22 @@ func resourceTFEProjectUpdate(ctx context.Context, d *schema.ResourceData, meta 
 	options := tfe.ProjectUpdateOptions{
 		Name:        tfe.String(d.Get("name").(string)),
 		Description: tfe.String(d.Get("description").(string)),
+	}
+
+	if tagBindings, ok := d.Get("tags").(map[string]interface{}); ok {
+		for key, val := range tagBindings {
+			options.TagBindings = append(options.TagBindings, &tfe.TagBinding{
+				Key:   key,
+				Value: val.(string),
+			})
+		}
+
+		if len(options.TagBindings) == 0 && !d.Get("ignore_additional_tags").(bool) {
+			err := config.Client.Projects.DeleteAllTagBindings(ctx, d.Id())
+			if err != nil {
+				return diag.Errorf("Error removing tag bindings from project %s: %v", d.Id(), err)
+			}
+		}
 	}
 
 	log.Printf("[DEBUG] Update configuration of project: %s", d.Id())