+++ internal/provider/resource_tfe_workspace_test.go
@@ -1252,6 +1252,50 @@ func TestAccTFEWorkspace_patternsAndPrefixesConflicting(t *testing.T) {
 	})
 }
 
+func TestAccTFEWorkspace_tagBindings(t *testing.T) {
+	skipUnlessBeta(t)
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccTFEWorkspace_basicTagBindings(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(
+						"tfe_worskpace.foobar", "tags.%", "3"),
+					resource.TestCheckResourceAttr(
+						"tfe_worskpace.foobar", "tags.keyA", "valueA"),
+					resource.TestCheckResourceAttr(
+						"tfe_worskpace.foobar", "tags.keyB", "valueB"),
+					resource.TestCheckResourceAttr(
+						"tfe_worskpace.foobar", "tags.keyC", "valueC"),
+				),
+			},
+			{
+				Config: testAccTFEWorkspace_basicTagBindingsRemoveOne(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(
+						"tfe_worskpace.foobar", "tags.%", "2"),
+					resource.TestCheckResourceAttr(
+						"tfe_worskpace.foobar", "tags.keyA", "valueA"),
+					resource.TestCheckResourceAttr(
+						"tfe_worskpace.foobar", "tags.keyB", "valueB"),
+				),
+			},
+			{
+				Config: testAccTFEWorkspace_basicTagBindingsRemoveAll(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(
+						"tfe_worskpace.foobar", "tags.%", "0"),
+				),
+			},
+		},
+	})
+}
+
 func TestAccTFEWorkspace_changeTags(t *testing.T) {
 	workspace := &tfe.Workspace{}
 	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
@@ -1266,7 +1310,7 @@ func TestAccTFEWorkspace_changeTags(t *testing.T) {
 		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
 		Steps: []resource.TestStep{
 			{
-				// create with 2 tags
+				// create with 2 legacy tags and 3 tag bindings
 				Config: testAccTFEWorkspace_basic(rInt),
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckTFEWorkspaceExists(
@@ -1408,6 +1452,38 @@ func TestAccTFEWorkspace_changeTags(t *testing.T) {
 						}
 						return fmt.Errorf("unmanaged tag not found on workspace")
 					},
+				),
+			},
+		},
+	})
+}
+
+func TestAccTFEWorkspace_effectiveTags(t *testing.T) {
+	skipUnlessBeta(t)
+
+	workspace := &tfe.Workspace{}
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckTFEWorkspaceDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccTFEWorkspace_effectiveTags(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckTFEWorkspaceExists(
+						"tfe_workspace.foobar", workspace, testAccProvider),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace.foobar", "tags.%", "1"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace.foobar", "effective_tags.%", "3"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace.foobar", "effective_tags.keyA", "valueA"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace.foobar", "effective_tags.keyB", "override"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace.foobar", "effective_tags.keyC", "valueC"),
 				),
 			},
 		},
@@ -2740,6 +2816,8 @@ func TestAccTFEWorkspace_updateWithAutoDestroyAt(t *testing.T) {
 }
 
 func TestAccTFEWorkspace_createWithAutoDestroyDuration(t *testing.T) {
+	skipUnlessAfterDate(t, time.Date(2025, time.February, 16, 0, 0, 0, 0, time.UTC))
+
 	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
 
 	resource.Test(t, resource.TestCase{
@@ -2759,6 +2837,8 @@ func TestAccTFEWorkspace_createWithAutoDestroyDuration(t *testing.T) {
 }
 
 func TestAccTFEWorkspace_updateWithAutoDestroyDuration(t *testing.T) {
+	skipUnlessAfterDate(t, time.Date(2025, time.February, 16, 0, 0, 0, 0, time.UTC))
+
 	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
 
 	resource.Test(t, resource.TestCase{
@@ -2892,6 +2972,67 @@ resource "tfe_workspace" "foobar" {
   tag_names          = ["fav", "test"]
   %s
 }`, rInt, aart)
+}
+
+func testAccTFEWorkspace_basicTagBindings(rInt int) string {
+	return fmt.Sprintf(`
+resource "tfe_organization" "foobar" {
+  name  = "tst-terraform-%d"
+  email = "admin@company.com"
+}
+
+resource "tfe_workspace" "foobar" {
+  name               = "workspace-test"
+  organization       = tfe_organization.foobar.id
+  description        = "My favorite workspace!"
+  allow_destroy_plan = false
+  auto_apply         = true
+  tag_names          = ["fav", "test"]
+  tags = {
+	  keyA = "valueA"
+	  keyB = "valueB"
+	  keyC = "valueC"
+  }
+}`, rInt)
+}
+
+func testAccTFEWorkspace_basicTagBindingsRemoveOne(rInt int) string {
+	return fmt.Sprintf(`
+resource "tfe_organization" "foobar" {
+  name  = "tst-terraform-%d"
+  email = "admin@company.com"
+}
+
+resource "tfe_workspace" "foobar" {
+  name               = "workspace-test"
+  organization       = tfe_organization.foobar.id
+  description        = "My favorite workspace!"
+  allow_destroy_plan = false
+  auto_apply         = true
+  tag_names          = ["fav", "test"]
+  tags = {
+	  keyA = "valueA"
+	  keyB = "valueB"
+  }
+}`, rInt)
+}
+
+func testAccTFEWorkspace_basicTagBindingsRemoveAll(rInt int) string {
+	return fmt.Sprintf(`
+resource "tfe_organization" "foobar" {
+  name  = "tst-terraform-%d"
+  email = "admin@company.com"
+}
+
+resource "tfe_workspace" "foobar" {
+  name               = "workspace-test"
+  organization       = tfe_organization.foobar.id
+  description        = "My favorite workspace!"
+  allow_destroy_plan = false
+  auto_apply         = true
+  tag_names          = ["fav", "test"]
+  tags = {}
+}`, rInt)
 }
 
 func testAccTFEWorkspace_HTMLURL(rInt int) string {
@@ -3307,7 +3448,7 @@ resource "tfe_workspace" "foobar" {
   name                  = "workspace-updated"
   organization          = tfe_organization.foobar.id
   description           = "My favorite workspace!"
-	assessments_enabled       = true
+  assessments_enabled   = true
   allow_destroy_plan    = false
   auto_apply            = true
   tag_names             = ["fav", "test"]
@@ -3924,5 +4065,32 @@ resource "tfe_workspace" "foobar" {
   tag_names          = ["fav", "test"]
   source_url         = "https://example.com"
   source_name        = "Example Source"
+}`, rInt)
+}
+
+func testAccTFEWorkspace_effectiveTags(rInt int) string {
+	return fmt.Sprintf(`
+resource "tfe_organization" "foobar" {
+  name  = "tst-terraform-%d"
+  email = "admin@company.com"
+}
+
+resource "tfe_project" "foobar" {
+  name = "project-test"
+  organization = tfe_organization.foobar.id
+  tags = {
+	  keyA = "valueA"
+	  keyB = "valueB"
+	  keyC = "valueC"
+  }
+}
+
+resource "tfe_workspace" "foobar" {
+  name               = "workspace-test"
+  organization       = tfe_organization.foobar.id
+  project_id         = tfe_project.foobar.id
+  tags               = {
+	  keyB = "override"
+  }
 }`, rInt)
 }